moved the tests to a dir
